pessoa

+++ Estac.Domain/Interface/Repositories/IBaseRepositories.cs
@@ -6,13 +6,13 @@ namespace Estac.Domain.Interface.Repositories
     {
         Task<T> Gravar(T item);
         Task<T> Alterar(T item);
-        Task<bool> Excluir(long id);
-        Task<T> Selecionar(long id);
+        Task<bool> Excluir(int id);
+        Task<T> Selecionar(int id);
         Task<IQueryable<T>> SelectAllAsync();
-        Task<bool> Existe(long id);
-        long LastCodeTable();
+        Task<bool> Existe(int id);
+        int LastCodeTable();
         Task<T> InsertNoSaveChangesAsync(T item);
         Task SaveChangesAsync();
-        Task<long?> GetIdByDescricaoAsync(string descricao);
+        Task<int?> GetIdByDescricaoAsync(string descricao);
     }
 }
